Split chat sidebar group rows
