+++ src/app/components/sidebar/chat/SidebarChatSection.tsx
@@ -2,43 +2,25 @@ import {
   DndContext,
   type DragEndEvent,
   type DragStartEvent,
-  type DraggableAttributes,
-  type DraggableSyntheticListeners,
   PointerSensor,
   closestCorners,
   useSensor,
   useSensors,
 } from "@dnd-kit/core";
 import { restrictToParentElement, restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import {
-  SortableContext,
-  arrayMove,
-  useSortable,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
-import {
-  ChevronDown,
-  ChevronRight,
-  Edit2,
-  Folder,
-  FolderPlus,
-  MoreHorizontal,
-  Plus,
-  Search,
-} from "lucide-react";
-import { type DragEvent, type ReactNode, useEffect, useMemo, useRef, useState } from "react";
+import { SortableContext, arrayMove, verticalListSortingStrategy } from "@dnd-kit/sortable";
+import { Edit2, FolderPlus, Plus, Search } from "lucide-react";
+import { type DragEvent, useMemo, useState } from "react";
 import type {
   ChatGroup,
   ChatSidebarState,
   ChatThread,
   DesktopActionInvoker,
 } from "../../../desktop/types";
-import { compactIconButtonClass } from "../../../ui/classes";
-import { cn } from "../../../utils/cn";
 import { SurfacePanel } from "../../common/SurfacePanel";
-import { Tooltip } from "../../common/Tooltip";
 import { IconButton } from "../../common/IconButton";
+import { ChatGroupRow } from "./ChatGroupRow";
+import { SortableGroupItem } from "./SortableGroupItem";
 import { ThreadRow } from "../project-tree/ThreadRow";
 import { useProjectMenuDismiss } from "../project-tree/useProjectMenuDismiss";
 
@@ -53,212 +35,6 @@ type SidebarChatSectionProps = {
   onRefresh: () => Promise<unknown>;
   onAction: DesktopActionInvoker;
 };
-
-type SortableGroupItemProps = {
-  groupId: string;
-  children: (input: {
-    dragHandleProps: {
-      attributes: DraggableAttributes;
-      listeners: DraggableSyntheticListeners | undefined;
-    };
-    isDragging: boolean;
-  }) => ReactNode;
-};
-
-function SortableGroupItem({ groupId, children }: SortableGroupItemProps) {
-  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
-    id: groupId,
-  });
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={{ transform: CSS.Translate.toString(transform), transition }}
-      className={isDragging ? "z-20 opacity-80" : undefined}
-    >
-      {children({ dragHandleProps: { attributes, listeners }, isDragging })}
-    </div>
-  );
-}
-
-type ChatGroupRowProps = {
-  actionMenuId: string;
-  actionMenuOpen: boolean;
-  dragHandleProps?: {
-    attributes: DraggableAttributes;
-    listeners: DraggableSyntheticListeners | undefined;
-  };
-  isActive: boolean;
-  isDragging: boolean;
-  isExpanded: boolean;
-  name: string;
-  renameDraft: string;
-  isEditing: boolean;
-  threadGroupId: string;
-  onCancelEdit: () => void;
-  onChangeRenameDraft: (value: string) => void;
-  onCreateSession: () => void;
-  onEdit: () => void;
-  onSelect: () => void;
-  onSubmitEdit: () => void;
-  onToggleActions: () => void;
-  onToggleExpanded: () => void;
-};
-
-function ChatGroupRow({
-  actionMenuId,
-  actionMenuOpen,
-  dragHandleProps,
-  isActive,
-  isDragging,
-  isExpanded,
-  name,
-  renameDraft,
-  isEditing,
-  threadGroupId,
-  onCancelEdit,
-  onChangeRenameDraft,
-  onCreateSession,
-  onEdit,
-  onSelect,
-  onSubmitEdit,
-  onToggleActions,
-  onToggleExpanded,
-}: ChatGroupRowProps) {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const clickTimeoutRef = useRef<number | null>(null);
-
-  useEffect(() => {
-    if (!isEditing) return;
-    inputRef.current?.focus();
-    inputRef.current?.select();
-  }, [isEditing]);
-
-  useEffect(() => {
-    return () => {
-      if (clickTimeoutRef.current !== null) window.clearTimeout(clickTimeoutRef.current);
-    };
-  }, []);
-
-  const handleRowClick = () => {
-    if (clickTimeoutRef.current !== null) window.clearTimeout(clickTimeoutRef.current);
-    clickTimeoutRef.current = window.setTimeout(() => {
-      onSelect();
-      onToggleExpanded();
-      clickTimeoutRef.current = null;
-    }, 180);
-  };
-
-  const handleRowDoubleClick = () => {
-    if (clickTimeoutRef.current !== null) {
-      window.clearTimeout(clickTimeoutRef.current);
-      clickTimeoutRef.current = null;
-    }
-    onEdit();
-  };
-
-  return (
-    <div
-      className="sidebar-row-surface sidebar-project-row"
-      data-highlighted={isActive || actionMenuOpen ? "true" : "false"}
-      data-dragging={isDragging ? "true" : "false"}
-    >
-      <Tooltip content={isExpanded ? "Collapse group" : "Expand group"} placement="right">
-        <button
-          type="button"
-          className="sidebar-project-toggle"
-          onClick={onToggleExpanded}
-          data-can-toggle="true"
-          aria-label={isExpanded ? "Collapse group" : "Expand group"}
-          aria-expanded={isExpanded}
-          aria-controls={threadGroupId}
-        >
-          <Folder size={12} className="sidebar-project-icon sidebar-project-origin-icon" />
-          {isExpanded ? (
-            <ChevronDown size={12} className="sidebar-project-icon sidebar-project-chevron-icon" />
-          ) : (
-            <ChevronRight size={12} className="sidebar-project-icon sidebar-project-chevron-icon" />
-          )}
-        </button>
-      </Tooltip>
-
-      {isEditing ? (
-        <div className="sidebar-project-edit">
-          <input
-            ref={inputRef}
-            value={renameDraft}
-            onChange={(event) => onChangeRenameDraft(event.target.value)}
-            onBlur={onCancelEdit}
-            onKeyDown={(event) => {
-              if (event.key === "Enter") {
-                event.preventDefault();
-                onSubmitEdit();
-              }
-              if (event.key === "Escape") {
-                event.preventDefault();
-                onCancelEdit();
-              }
-            }}
-            className="sidebar-project-input"
-            aria-label={`Rename ${name}`}
-          />
-        </div>
-      ) : (
-        <button
-          type="button"
-          className={cn(
-            "sidebar-project-button",
-            dragHandleProps ? "cursor-grab active:cursor-grabbing" : "cursor-pointer",
-          )}
-          {...dragHandleProps?.attributes}
-          {...dragHandleProps?.listeners}
-          onClick={handleRowClick}
-          onDoubleClick={handleRowDoubleClick}
-          data-active={isActive ? "true" : "false"}
-          aria-current={isActive ? "page" : undefined}
-        >
-          <span className="sidebar-project-title">{name}</span>
-        </button>
-      )}
-
-      <div
-        className="sidebar-project-actions"
-        data-open={actionMenuOpen ? "true" : "false"}
-        data-dragging={isDragging ? "true" : "false"}
-        data-editing={isEditing ? "true" : "false"}
-        data-visible="true"
-      >
-        <Tooltip content="New chat" placement="right">
-          <button
-            type="button"
-            className={compactIconButtonClass}
-            onClick={onCreateSession}
-            aria-label={`Start a new chat in ${name}`}
-          >
-            <Plus size={14} />
-          </button>
-        </Tooltip>
-
-        <Tooltip content="Group actions" placement="right">
-          <button
-            type="button"
-            className={cn(
-              compactIconButtonClass,
-              actionMenuOpen && "bg-[rgba(255,255,255,0.05)] text-[color:var(--text)]",
-            )}
-            onClick={onToggleActions}
-            aria-label="Group actions"
-            aria-haspopup="menu"
-            aria-expanded={actionMenuOpen}
-            aria-controls={actionMenuId}
-          >
-            <MoreHorizontal size={14} />
-          </button>
-        </Tooltip>
-      </div>
-    </div>
-  );
-}
 
 export function SidebarChatSection({
   chatState,
